Move auditor classes into separate files.
